GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - src/datasets/subs/utxos_metadata.rs
   - src/structs/parse/date_map.rs
   - src/datasets/address/structs.rs
   - src/datasets/mod.rs
   - src/datasets/address/mod.rs
   - src/structs/parse/database.rs
   - src/datasets/address/dataset.rs
   - src/datasets/address/datasets.rs
   - src/databases/mod.rs
   - src/min_height.rs
   - src/datasets/utxo/datasets.rs
   - src/structs/parse/height_map.rs
   - src/datasets/_traits.rs
   - src/datasets/utxo/dataset.rs

Your previous branch was: refs/heads/main

The sha for that commit was: 28dfe4373803a5f15b7962de59ff175babffdc22

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch